fix float

+++ parse_new/main3.cpp
@@ -50,11 +50,11 @@ int main(int argc, const char * argv[]) {
     // find 99th percentile
     unsigned int n99 = latency.size() * 0.99 - 1;
     cout << "99th index: " << n99 << endl;
-    unsigned int p99 = latency[n99];
+    float p99 = latency[n99];
     cout << "99th percentile latency: " << p99 << endl;
     unsigned int n9999 = latency.size() * 0.9999 - 1;
     cout << "9999th index: " << n9999 << endl;
-    unsigned int p9999 = latency[n9999];
+    float p9999 = latency[n9999];
     cout << "99.99th percentile latency: " << p9999 << endl;
     return 0;
 }
